Fix average db display
Fixes #357
